refactor: extract state apply mixin from coordinator core
